add dom helpers to break up index.js
